refactor(frontend): restructure workspace routes
