--- frontend/src/app/routes.ts
@@ -1,3 +1,0 @@
-import { flatRoutes } from "@react-router/fs-routes";
-
-export default flatRoutes();
