+++ frontend/src/app/routes/workspaces/$workspaceId/_index.tsx
