+++ frontend/src/app/routes/workspaces/$workspaceId/configurations/_index.tsx
