+++ frontend/src/app/routes/workspaces/$workspaceId/documents/$documentId/route.tsx
