+++ frontend/src/app/routes/workspaces/$workspaceId/documents/_index.tsx
