+++ frontend/src/app/routes/workspaces/$workspaceId/jobs/route.tsx
@@ -1,0 +1,5 @@
+import { Outlet } from "react-router-dom";
+
+export default function WorkspaceJobsLayoutRoute() {
+  return <Outlet />;
+}
