+++ frontend/src/app/routes/workspaces/$workspaceId/settings/route.tsx
@@ -1,0 +1,5 @@
+import { Outlet } from "react-router-dom";
+
+export default function WorkspaceSettingsLayoutRoute() {
+  return <Outlet />;
+}
