+++ frontend/src/app/routes/workspaces/route.tsx
@@ -1,0 +1,5 @@
+import { Outlet } from "react-router-dom";
+
+export default function WorkspacesLayoutRoute() {
+  return <Outlet />;
+}
