+++ frontend/src/features/workspaces/components/GlobalTopBar.tsx
