+++ frontend/src/features/workspaces/lib/navigation.tsx
@@ -1,4 +1,6 @@
-import type { SVGProps } from "react";
+import type { ComponentType, SVGProps } from "react";
+
+import type { WorkspaceProfile } from "@types/workspaces";
 
 interface IconProps extends SVGProps<SVGSVGElement> {
   readonly title?: string;
@@ -40,3 +42,44 @@ export const SettingsIcon = createIcon(
 export const DirectoryIcon = createIcon(
   "M4 10.5 12 4l8 6.5V19a1 1 0 0 1-1 1h-4v-5h-6v5H5a1 1 0 0 1-1-1v-8.5Z",
 );
+
+export type WorkspaceSectionId = "documents" | "configurations" | "settings";
+
+interface WorkspaceSectionDescriptor {
+  readonly id: WorkspaceSectionId;
+  readonly path: string;
+  readonly label: string;
+  readonly icon: ComponentType<SVGProps<SVGSVGElement>>;
+}
+
+const workspaceSections: readonly WorkspaceSectionDescriptor[] = [
+  {
+    id: "documents",
+    path: "documents",
+    label: "Documents",
+    icon: DocumentsIcon,
+  },
+  {
+    id: "configurations",
+    path: "configurations",
+    label: "Configurations",
+    icon: ConfigureIcon,
+  },
+  {
+    id: "settings",
+    path: "settings",
+    label: "Workspace Settings",
+    icon: SettingsIcon,
+  },
+] as const;
+
+export const defaultWorkspaceSection = workspaceSections[0];
+
+export function getWorkspacePrimaryNavigation(workspace: WorkspaceProfile) {
+  return workspaceSections.map((section) => ({
+    id: section.id,
+    label: section.label,
+    href: `/workspaces/${workspace.id}/${section.path}`,
+    icon: section.icon,
+  }));
+}
